+++ src/App.vue
@@ -3,7 +3,8 @@
     <div id="nav">
       <router-link to="/">Home</router-link> |
       <router-link to="/about">About</router-link> |
-      <router-link to="/author">Write</router-link>
+      <router-link to="/author">Write</router-link> |
+      <router-link to="/login">Login</router-link>
     </div>
     <router-view/>
   </div>
